+++ pkg/food/display.go
@@ -2,16 +2,18 @@ package food
 
 // SimplifiedStyle is a simplified version of style for display
 type SimplifiedStyle struct {
-	CommonDishes []string `json:"common_dishes"`
-	EatingTraits []string `json:"eating_traits"`
-	Breads       []string `json:"breads"`
+	CommonDesserts   []string `json:"common_desserts"`
+	CommonMainDishes []string `json:"common_main_dishes"`
+	EatingTraits     []string `json:"eating_traits"`
+	Breads           []string `json:"breads"`
 }
 
 // Simplify returns a simplified style
 func (style Style) Simplify() SimplifiedStyle {
 	return SimplifiedStyle{
-		CommonDishes: style.CommonDishes,
-		EatingTraits: style.EatingTraits,
-		Breads:       style.Breads,
+		CommonDesserts:   style.CommonDesserts,
+		CommonMainDishes: style.CommonMainDishes,
+		EatingTraits:     style.EatingTraits,
+		Breads:           style.Breads,
 	}
 }