+++ pkg/food/styles.go
@@ -9,13 +9,16 @@ import (
 
 // Style is a cultural food style
 type Style struct {
-	Breads            []string
-	CommonBases       []string
-	CommonDishes      []string
-	CommonSpices      []string
-	CookingTechniques []string
-	EatingTraits      []string
-	Noodles           []string
+	Breads             []string
+	CommonBases        []string
+	CommonDessertBases []string
+	CommonDesserts     []string
+	CommonMainDishes   []string
+	CommonSpices       []string
+	CookingTechniques  []string
+	EatingTraits       []string
+	DessertTypes       []string
+	Noodles            []string
 }
 
 // GenerateStyle procedurally generates a style of food
@@ -35,12 +38,33 @@ func GenerateStyle(originClimate climate.Climate) Style {
 			} else {
 				style.CommonSpices = append(style.CommonSpices, r.Name)
 			}
-		} else if r.Type == "fruit" {
+		} else if r.Type == "egg" {
 			style.CommonBases = append(style.CommonBases, r.Name)
+		} else if r.Type == "milk" {
+			if !slices.StringIn("pudding", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "pudding")
+			}
+		} else if r.Type == "fruit" {
+			style.CommonDessertBases = append(style.CommonDessertBases, r.Name)
+			if !slices.StringIn("fruit", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "fruit")
+			}
 		} else if r.Type == "vegetable" {
 			style.CommonBases = append(style.CommonBases, r.Name)
 		} else if r.Type == "flour" {
 			style.Noodles = append(style.Noodles, getNoodles(r.Name)...)
+			if !slices.StringIn("pie", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "pie")
+			}
+			if !slices.StringIn("crisp", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "crisp")
+			}
+			if !slices.StringIn("tart", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "tart")
+			}
+			if !slices.StringIn("cake", style.DessertTypes) {
+				style.DessertTypes = append(style.DessertTypes, "cake")
+			}
 		}
 	}
 
@@ -50,7 +74,8 @@ func GenerateStyle(originClimate climate.Climate) Style {
 
 	style.CookingTechniques = randomTechniques(3)
 
-	style.CommonDishes = style.randomDishes()
+	style.CommonDesserts = style.randomDesserts()
+	style.CommonMainDishes = style.randomMainDishes()
 	style.Breads = randomBreads(originClimate)
 	style.EatingTraits = randomEatingTraits()