+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.3.2 (2019-07-19)
+
+Changes to food.
+
+### Change List
+
+- Add basic desserts
+- Remove fruit from main dishes as a base
+
 ## 0.3.1 (2019-07-19)
 
 Small bug fixes.